+++ AudioVisio/src/audiovisio/level/Stair.java
@@ -21,9 +21,9 @@ public class Stair extends Panel{
     
     public static Stair load(JSONObject obj) {
     	JSONObject location = (JSONObject) obj.get("location");
-    	float locX = (float) location.get("x");
-    	float locY = (float) location.get("y");
-    	float locZ = (float) location.get("z");
+    	float locX = ((Double)location.get("x")).floatValue();
+    	float locY = ((Double)location.get("y")).floatValue();
+    	float locZ = ((Double)location.get("z")).floatValue();
     	
     	Vector3f locVec = new Vector3f(locX, locY, locZ);
